Fixed all the routing errors, moved components back into /src, and commented out code that isnt functional yet
