+++ client/src/pages/paymentSuccess.js
@@ -1,0 +1,5 @@
+// function Completion(props) {
+//     return <h1>Thank you!</h1>
+// }
+
+// export default PaymentSuccess;
