--- client/src/pages/paymentSuccessful.js
@@ -1,5 +1,0 @@
-function Completion(props) {
-    return <h1>Thank you!</h1>
-}
-
-export default Completion;
